+++ flake.nix
@@ -110,6 +110,7 @@ rec {
             exec ${pkgs.python3}/bin/python3 "$@"
           '';
           buck2-archive-tool = buck2-stage0-tools.archive-tool;
+          buck2-events = buck2-stage0-tools.events;
           buck2-product = buck2-stage0-tools.product;
           buck2-coreutils = pkgs.writeShellScriptBin "readlink" ''
             exec ${pkgs.coreutils}/bin/readlink "$@"